requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.48.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
six==1.16.0

requests==2.31.0
pytz==2024.1
arrow==1.3.0
maya==0.6.1
SQLAlchemy==2.0.27
python-dotenv==1.0.1
paramiko==3.4.0
pexpect==4.9.0
pydantic==2.6.2
fastapi==0.110.0
uvicorn==0.27.1